api/pyproject.toml
[project]
name = "gabi-api"
version = "0.1.0"
description = "Gabi Hub — Unified AI Backend (nGhost + Law & Comply + nTalkSQL)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "pymssql>=2.2.0",
    "sentence-transformers>=3.3.0",
    "numpy>=1.26.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-auth>=2.35.0",
    "firebase-admin>=6.6.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.18",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
